workout_tracker/pyproject.toml
[project]
name = "workout-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.1.6",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.4.0",
    "setuptools>=78.1.0",
    "wheel>=0.45.1",
]

[project.optional-dependencies]
types-stubs = [
    "django-stubs>=5.1.3",
    "django-types>=0.20.0",
    "djangorestframework-stubs>=3.15.3",
    "djangorestframework-types>=0.9.0",
]


[dependency-groups]
dev = [
    "ipython>=9.0.1",
    "pytest-django>=4.10.0",
    "ruff>=0.9.7",
]

[tool.ruff.lint]
exclude = [".venv"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "workout_tracker.settings"
python_files = ["test_*.py"]
addopts = "--verbose -rP --log-cli-level=INFO"